+++ yandex/data_source_yandex_compute_instance_group.go
@@ -690,6 +690,42 @@ func dataSourceYandexComputeInstanceGroup() *schema.Resource {
 				},
 			},
 
+			"application_load_balancer": {
+				Type:     schema.TypeList,
+				MaxItems: 1,
+				Computed: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"target_group_name": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"target_group_description": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"target_group_labels": {
+							Type:     schema.TypeMap,
+							Computed: true,
+							Elem:     &schema.Schema{Type: schema.TypeString},
+							Set:      schema.HashString,
+						},
+						"target_group_id": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"status_message": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"max_opening_traffic_duration": {
+							Type:     schema.TypeInt,
+							Computed: true,
+						},
+					},
+				},
+			},
+
 			"created_at": {
 				Type:     schema.TypeString,
 				Computed: true,
@@ -798,6 +834,24 @@ func dataSourceYandexComputeInstanceGroup() *schema.Resource {
 				},
 			},
 
+			"application_balancer_state": {
+				Type:     schema.TypeList,
+				MaxItems: 1,
+				Computed: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"target_group_id": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"status_message": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+					},
+				},
+			},
+
 			"status": {
 				Type:     schema.TypeString,
 				Computed: true,
@@ -853,6 +907,14 @@ func flattenInstanceGroupDataSource(d *schema.ResourceData, instanceGroup *insta
 		return err
 	}
 	if err := d.Set("load_balancer_state", loadBalancerState); err != nil {
+		return err
+	}
+
+	applicationLoadBalancerState, err := flattenInstanceGroupApplicationLoadBalancerState(instanceGroup)
+	if err != nil {
+		return err
+	}
+	if err := d.Set("application_load_balancer", applicationLoadBalancerState); err != nil {
 		return err
 	}